test: fix pmem_movnt_align to properly work on ppc64
This test was timing out on ppc64 for two main reasons. First of all, it
tests sizes multiples of the page size which is 16 times bigger in
ppc64. Besides this, the x86_64 version already have some optimized
functions for this. Also the both existing tests use x86_64 specific
environment variables.

This commit adds restrict x86_64 requirement for TEST0 and TEST1. It also
creates a third test to properly work on ppc64.