test: fix obj_heap_state for ppc64
As the header's size change for different page sizes the
objects offset also change. This makes the test fail for
page sizes other than 4kB.

Other factor is the cacheline size that will affect the
allocation sizes.

This fix adds a calculation to normalize the oid number.
This way it will pass for any different page size.

Signed-off-by: Lucas A. M. Magalhaes <lamm@linux.ibm.com>